Update to compile and run on more modern versions of FreeBSD.
Many thanks to Miles Lott for preliminary work for this update.

The major change is the complete replacement to the configuration code for
the tunnel interface and the routes.  The code now simply calls ifconfig(8)
and route(8) with the appropriate paramters, instead of using the ioctl and
PF_ROUTE APIs.

The remainder of the changes adjust the data types and the struct alignment.